create admin controller and move relevant methods in
